requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
google-generativeai
